requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
pandas
numpy
scikit-learn
xgboost
requests
beautifulsoup4
python-dotenv
pydantic-settings
neo4j
networkx
pyvis
redis
pytest
loguru
